+++ src/components/fridgeList/InputDelete.tsx
@@ -50,8 +50,8 @@ const CategoreDelete = ({ onDeleteCategory }: { onDeleteCategory: (keywords: str
         onKeyDown={handleKeyDown}
         onCompositionStart={handleComposition}
         onCompositionEnd={handleComposition}
-        placeholder="빼고 싶은 재료를 입력하세요 XD"
-        className="border p-1"
+        placeholder="빼고 싶은 재료를 입력하세요!"
+        className="border"
       />
       <button type="button" onClick={addCategory} className="ml-2 p-1 border">
         추가
